Conflicts:
	core/sj-crud-rest/src/main/scala/com/bwsw/sj/crud/rest/validator/SjCrudValidator.scala
	core/sj-output-streaming-engine/src/main/scala/com/bwsw/sj/engine/output/subscriber/OutputSubscriberCallback.scala
	core/sj-regular-streaming-engine/src/main/scala/com/bwsw/sj/engine/regular/subscriber/RegularConsumerCallback.scala
	core/sj-regular-streaming-engine/src/main/scala/com/bwsw/sj/engine/regular/task/engine/input/KafkaTaskInputService.scala
	core/sj-regular-streaming-engine/src/main/scala/com/bwsw/sj/engine/regular/task/engine/input/RegularTaskInputService.scala

+++ core/sj-common/src/main/scala/com/bwsw/common/tstream/NetworkTimeUUIDGenerator.scala
@@ -17,7 +17,7 @@ import org.slf4j.LoggerFactory
 
 class NetworkTimeUUIDGenerator(zkServers: Array[String],
                               prefix: String,
-                              retryInterval: Long,
+                              retryInterval: Int,
                               retryCount: Int) extends IUUIDGenerator {
 
   private val logger = LoggerFactory.getLogger(this.getClass)
